meaning_unit_extractor/requirements.txt
# 运行时必备
pyyaml==6.0.2
requests==2.32.3
pandas==2.2.3
python-docx==1.1.2

# 测试
pytest==8.3.4

# 可选（semantic_refine 启用时需要；默认不启用）
# sentence-transformers==3.1.1
# torch>=2.0
